Break dash into modules
