+++ src/lib/dash/args.js
@@ -1,0 +1,7 @@
+export const parseArg = (arg) =>
+  arg.includes(",") ? arg.split(",").map((item) => item.trim()) : arg || null;
+
+export const validateArg = (arg) =>
+  Array.isArray(arg)
+    ? arg.every(validateArg)
+    : arg === null || arg.match(/^0x[0-9a-f]{1,32}$/i);
