+++ src/lib/dash/random.js
@@ -1,0 +1,9 @@
+export const getRandomBase64 = (length = 32) =>
+  btoa(
+    String.fromCharCode.apply(
+      null,
+      typeof window === "undefined"
+        ? []
+        : crypto.getRandomValues(new Uint8Array(length))
+    )
+  );
